2nd commit

+++ src/main/java/com/example/contact/controller/UserController.java
@@ -6,6 +6,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import javax.validation.Valid;
+
 @RestController
 @RequestMapping("/users")
 @CrossOrigin(origins = "")
@@ -21,22 +23,34 @@ public class UserController {
 
     @GetMapping("/get_id/{id}")
     public ResponseEntity<?> getUserById(@PathVariable("id") Integer id) {
+        if (id == null) {
+            throw new RuntimeException("id cannot be null");
+        }
         return ResponseEntity.ok(userService.getById(id));
     }
 
-
     @PostMapping("/create")
-    public ResponseEntity<?> createUser(@RequestBody User user) {
+    public ResponseEntity<?> createUser(@Valid @RequestBody User user) {
+        if (user == null) {
+            throw new RuntimeException("user cannot be null");
+        }
         return ResponseEntity.ok(userService.createUser(user));
     }
 
     @PutMapping("/update/{id}")
     public ResponseEntity<?> updateUser(@RequestBody User user, @PathVariable("id") Integer id) {
+        if (user == null) {
+            throw new RuntimeException("User cannot be null");
+        }
+        if (id == null) throw new RuntimeException("id cannot be null");
         return ResponseEntity.ok(userService.updateUser(user, id));
     }
 
     @DeleteMapping("/delete/{id}")
     public ResponseEntity<?> deleteUser(@PathVariable("id") Integer id) {
+        if (id == null) {
+            throw new RuntimeException("id cannot be null");
+        }
         return ResponseEntity.ok(userService.deleteUser(id));
     }
 }
